requirements.txt
pandas==1.5.3
python-dotenv==0.21.1
SQLAlchemy==2.0.4
mysql.connector
csv